+++ katarl2/agents/simba_ppo/simba_ppo.py
@@ -264,7 +264,7 @@ class SimbaPPO(BaseAgent):
             self.PATH_CKPTS = PATH_LOGS / "ckpts"
             self.PATH_CKPTS.mkdir(exist_ok=True, parents=True)
 
-            path_ckpt = self.PATH_CKPTS / f"sac-{self.train_step}.pkl"
+            path_ckpt = self.PATH_CKPTS / f"{self.cfg.full_name}-{self.train_step}.pkl"
         else:
             path_ckpt = path
         torch.save(data, str(path_ckpt))